0.0.6: Add support for multiple template source files.
Previously, this module was hardcoded to look for pre-compiled Handlebars templates
in `rendr.entryPath + '/app/templates/compiledTemplates'`. Now, that's the default,
but apps that use this module can manipulate the `exports.templatePatterns` array
to specify the template path patterns that match a given source file.

This is useful for when a Rendr app supports different sets of templates for different
pages; for example, an app could have a set of desktop templates and a set of
mobile templates.